dashboard sidebar refactored
