+++ frontend/app/globals.css
@@ -77,7 +77,8 @@
   --color-custom-gray-7:   #EAEAEA;
   --color-custom-gray-8:   #EBEBEB;
   --color-custom-gray-9:   #F3F4F6;
-  --color-custom-blue-gray-1: '#6E7E87'
+  --color-custom-gray-10: #D5D6D8;
+  --color-custom-blue-gray-1: #6E7E87;
 
   /* — Custom font token — */
   --font-montserrat: 'Montserrat', sans-serif;
